Chore: Move components directory
